feat: migrate to pnpm package manager
This commit migrates the project to use pnpm as the package manager.
This change was made to improve dependency management and installation
speed.

feat(Dockerfile): use pnpm for dependency installation

This commit updates the Dockerfile to use pnpm for installing
dependencies. This change aligns with the project's migration to pnpm
and ensures that the Docker image is built using the correct package
manager.

feat: update .gitignore to include .pnpm-store directory

This commit adds the `.pnpm-store` directory to the `.gitignore` file.
This directory is used by pnpm to store globally installed packages,
and it should not be committed to the repository.

fix: update prisma client and prisma version

This commit updates the prisma client and prisma version to the latest
version. This change was made to improve performance and stability.

remove: package 2.json
This commit removes the package 2.json file. This file is not needed
anymore.